% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfpolymorph-package.R
\docType{package}
\name{sfpolymorph-package}
\alias{sfpolymorph}
\alias{sfpolymorph-package}
\title{sfpolymorph: Calculate polygon shape metrics}
\description{
Calculate polygon shape metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsta/sfpolymorph}
  \item Report bugs at \url{https://github.com/jsta/sfpolymorph/issues}
}

}
\author{
\strong{Maintainer}: Joseph Stachelek \email{stachel2@msu.edu} (0000-0002-5924-2464)

}
\keyword{internal}
